Your data in a blink of an eye

blink summarizes your data tailored to the need of ophthalmic research: It looks for VA and IOP columns and summarises those with common statistics. In order to make it work, it requires specific column naming - please see section "column names" and "data coding". For more details how blink works, see vignette("eye")

Usage

blink(x, va_to = "logmar", va_cols = NULL, iop_cols = NULL, fct_level = 0:4)

Arguments

x

data frame

va_to

to which VA notation (passed to va())

va_cols

if specified, overruling automatic VA columns selection. tidyselection supported

iop_cols

if specified, overruling automatic IOP columns selection. tidyselection supported

fct_level

Remove columns for Summarizing when all unique values fall into range. character or numeric vector, default 1:4

Value

object of class blink and list. Class blink contains the myopized data, count of patients and eyes, and summaries for visual acuities and intraocular pressure.

Details

blink is basically a wrapper around myop, eyes and reveal:

Duplicate rows are always removed

Column names are prepared for myopization (see myop)

VA will always be converted to logmar

Data coding

Only common codes supported:

eyes: "r", "re", "od", "right" - or numeric coding r:l = 0:1 or 1:2

Visual acuity: "VA", "BCVA", "Acuity"

Intraocular pressure: "IOP", "GAT", "NCT", "pressure"

Column name rules

No spaces!

Do not use numeric coding for eyes in column names

Separate eye and VA and IOP codes with underscores ("bcva_l_preop", "VA_r", "left_va", "IOP_re")

Avoid separate VA or IOP codes if this is not actually containing VA/ IOP data (e.g. "stableVA" instead of "stable_va", ChangeIOP instead of "change_IOP")

Keep names short

Don't use underscores when you don't have to. Consider each section divided by an underscore as a relevant characteristic of your variable. ("preop" instead of "pre_op", "VA" instead of "VA_ETDRS_Letters")

Use common codes for your patient column (see eyes, section Guessing) (e.g., "pat", "patient" or "ID", ideally both: "patientID" or "patID")

Don't be too creative with your names!

Names examples

Good names:

-c("patid", "surgery_right", "iop_r_preop", "va_r_preop", "iop_r", "iop_l")

OK names

-c("Id", "Eye", "BaselineAge", "VA_ETDRS_Letters", "InjectionNumber"): Names are long and there are two unnecessary underscore in the VA column. Better just "VA" -c("id", "r", "l"): All names are commonly used (good!), but which dimension of "r"/"l" are we exactly looking at?

Bad names (eye will fail)

c("id", "iopr", "iopl", "VAr", "VAl"): eye won't be able to recognize IOP and VA columns

c("id", "iop_r", "iop_l", "stable_iop_r", "stable_iop_l"): eye may wrongly identify the (probably logical) columns "stable_iop" as columns containing IOP data. Better maybe: "stableIOP_l"

c("person", "goldmann", "vision"): eye will not recognize that at all

tidy data

blink and myop work more reliably with clean data (any package will, really!). clean data.

column removal

Done with remCols: Removes columns that only contain values defined in fct_levels or logicals from selected columns (currently for both automatically and manually selected columns). fct_levels are removed because they are likely categorical codes.

See also

Examples

library(eyedata)
blink(amd2)
#> The lifecycle of blink() has expired. It will no longer be
#>   maintained, but will be kept in the package.
#> Warning: Data seems already myopic - no changes made. ?myop for help
#> From etdrs. Could be etdrs, logmar, snellendec
#> 
#> ── blink ───────────────────────────────────────────────────────────────────────
#> ══ Data ════════════════════════════════
#> # A tibble: 40,764 × 7
#>    patid sex    age0 eye    time va       inj_no
#>    <chr> <chr> <dbl> <chr> <dbl> <logmar>  <dbl>
#>  1 id_1  f        99 l         0 0.82          1
#>  2 id_1  f        99 l        28 0.08          2
#>  3 id_1  f        99 l        67 0.70          3
#>  4 id_1  f        99 l       147 0.90          4
#>  5 id_1  f        99 l       259 1.06          5
#>  6 id_1  f        99 l       343 1.02          6
#>  7 id_1  f        99 l       399 0.96          7
#>  8 id_1  f        99 l       534 1.06          8
#>  9 id_2  m        82 r         0 0.40          1
#> 10 id_2  m        82 r        39 0.46          2
#> # ℹ 40,754 more rows
#> 
#> ══ Count of patient and eyes ═══════════
#> ══ Counts ═══════════════
#>    id  eyes right  left 
#>  3357  3357  1681  1676 
#> 
#> ══ Visual acuity ═══════════════════════
#> 
#> ── $VA_total (all eyes) 
#>   var mean  sd     n  min max
#> 1  va  0.5 0.3 40762 -0.3 2.3
#> 
#> ── $VA_eyes (right and left eyes) 
#>   var1 var2 var mean  sd     n  min max
#> 1    l    r  va  0.5 0.3 20458 -0.3 2.3
#> 2    l    r  va  0.5 0.3 20304 -0.3 2.3
#> 
#> ══ Intraocular pressure ════════════════
#> 
#> ── $IOP_total (all eyes) 
#> NULL
#> 
#> ── $IOP_eyes (right and left eyes) 
#> NULL

messy_df <- data.frame( id = letters[1:3],
iop_r_preop = sample(21:23), iop_r_postop = sample(11:13),
iop_l_postop = sample(11:13), iop_l_preop = sample(31:33),
va_r_preop = sample(41:43),  va_l_preop = sample(41:43),
va_r_postop = sample(51:53), va_l_postop = sample(45:47)
)
blink(messy_df)
#> The lifecycle of blink() has expired. It will no longer be
#>   maintained, but will be kept in the package.
#> From etdrs
#> From etdrs
#> 
#> ── blink ───────────────────────────────────────────────────────────────────────
#> ══ Data ════════════════════════════════
#> # A tibble: 6 × 6
#>   id    eye   iop_preop iop_postop va_preop va_postop
#>   <chr> <chr> <chr>     <chr>      <logmar> <logmar> 
#> 1 a     right 21        13         0.84     0.66     
#> 2 a     left  31        12         0.86     0.80     
#> 3 b     right 22        11         0.88     0.68     
#> 4 b     left  33        11         0.88     0.78     
#> 5 c     right 23        12         0.86     0.64     
#> 6 c     left  32        13         0.84     0.76     
#> 
#> ══ Count of patient and eyes ═══════════
#> ══ Counts ═══════════════
#>    id  eyes right  left 
#>     3     6     3     3 
#> 
#> ══ Visual acuity ═══════════════════════
#> 
#> ── $VA_total (all eyes) 
#>         var mean  sd n min max
#> 1  va_preop  0.9 0.0 6 0.8 0.9
#> 2 va_postop  0.7 0.1 6 0.6 0.8
#> 
#> ── $VA_eyes (right and left eyes) 
#>   var1 var2  var3  var4       var mean sd n min max
#> 1 left left right right  va_preop  0.9  0 3 0.8 0.9
#> 2 left left right right va_postop  0.8  0 3 0.8 0.8
#> 3 left left right right  va_preop  0.9  0 3 0.8 0.9
#> 4 left left right right va_postop  0.7  0 3 0.6 0.7
#> 
#> ══ Intraocular pressure ════════════════
#> 
#> ── $IOP_total (all eyes) 
#>          var mean  sd n min max
#> 1  iop_preop   27 5.5 6  21  33
#> 2 iop_postop   12 0.9 6  11  13
#> 
#> ── $IOP_eyes (right and left eyes) 
#>   var1 var2  var3  var4        var mean sd n min max
#> 1 left left right right  iop_preop   32  1 3  31  33
#> 2 left left right right iop_postop   12  1 3  11  13
#> 3 left left right right  iop_preop   22  1 3  21  23
#> 4 left left right right iop_postop   12  1 3  11  13