reveal

Shows commonly used summary statistics

Usage

reveal(x, by = NULL, dec = 1, funs = NULL)

Arguments

x

data frame, numeric vector, or list of numeric vectors

by

character vector with the names of the columns. Can be several variables!

dec

how many decimals are displayed

funs

not really meant to be used at the moment - change the Summarizing functions with a named(!) list of functions

Value

data frame

Details

Character vectors (or character columns) will be removed.

See also

Examples

x = y = z = c(rnorm(20), NA)
mylist <- list(x = x, y = y, z = z)
## vectors
reveal(x)
#> mean   sd    n  min  max 
#>  0.0  0.8 20.0 -2.0  1.4 
reveal(1:10)
#> mean   sd    n  min  max 
#>  5.5  3.0 10.0  1.0 10.0 
## named or unnamed list
reveal(mylist)
#>   var1 var2 var3 mean  sd  n min max
#> 1    x    y    z    0 0.8 20  -2 1.4
#> 2    x    y    z    0 0.8 20  -2 1.4
#> 3    x    y    z    0 0.8 20  -2 1.4
set.seed(42)
mydf <- cbind(group = rep(letters[1:3], 4),
setNames(as.data.frame(replicate(c(rnorm(11), NA), n = 3)), letters[24:26]))
## data frames
reveal(mydf)
#> Warning: reveal: character columns removed
#>   var mean  sd  n  min max
#> 1   x  0.6 0.8 11 -0.6 2.0
#> 2   y -0.5 1.5 11 -2.7 2.3
#> 3   z  0.2 1.0 11 -1.8 1.9
## data frames by group
reveal(mydf, by = "group")
#>   var1 var2 var3 var4 var5 var6 var7 var8 var9 var mean  sd n  min max
#> 1    a    a    a    b    b    b    c    c    c   x  0.9 0.7 4 -0.1 1.5
#> 2    a    a    a    b    b    b    c    c    c   y -0.2 2.0 4 -2.7 2.3
#> 3    a    a    a    b    b    b    c    c    c   z  0.1 0.5 4 -0.4 0.7
#> 4    a    a    a    b    b    b    c    c    c   x  0.3 0.8 4 -0.6 1.3
#> 5    a    a    a    b    b    b    c    c    c   y -1.2 1.3 4 -2.4 0.6
#> 6    a    a    a    b    b    b    c    c    c   z  0.3 0.9 4 -0.6 1.2
#> 7    a    a    a    b    b    b    c    c    c   x  0.8 1.1 3 -0.1 2.0
#> 8    a    a    a    b    b    b    c    c    c   y  0.3 0.9 3 -0.3 1.3
#> 9    a    a    a    b    b    b    c    c    c   z  0.2 1.8 3 -1.8 1.9