VA conversion methods

S3 methods for VA conversion

Usage

convertVA(x, ...)

# S3 method for class 'quali'
convertVA(x, to, type, ...)

# S3 method for class 'snellendec'
convertVA(x, to, type, ...)

# S3 method for class 'snellen'
convertVA(x, to, type, smallstep, noplus, ...)

# S3 method for class 'logmar'
convertVA(x, to, type, ...)

# S3 method for class 'etdrs'
convertVA(x, to, type, ...)

# Default S3 method
convertVA(x, to, ...)

Arguments

x

vector of visual acuities

...

further arguments passed to methods

to

to which VA class to convert

type

which snellen notation. One of "ft", "m" or "dec"

smallstep

how plus/minus entries are evaluated. Default to increase/decrease snellen fractions by lines. If TRUE, each snellen optotype will be considered equivalent to 0.02 logmar or 1 ETDRS letter (assuming 5 letters in a row in a chart)

noplus

ignoring plus/minus entries and just returning the snellen fraction. This overrides the smallstep argument.

Value

vector with visual acuity of class va. See also "VA classes"

Details

VA can be snellen feet/meter/decimal, logMAR, ETDRS, or "qualitative" (Counting fingers, etc.)

Snellen fractions need to be either form 6/x or 20/x

ETDRS must be between 0 and 100

logMAR must be between -0.3 and 3.0

Qualitative must be PL, LP, NLP, NPL, HM, CF (any case allowed)

Any element which is not recognized will be converted to NA

Conversion

For other conversion and theory behind conversion rules see va section VA conversion.

plus/minus entries

The following rules for plus minus notations will be applied:

if entry -2 to +2 : take same Snellen value

if < -2 : take Snellen value one line below

if > +2: Snellen value one line above

Snellen are unfortunately often entered with "+/-", which is a violation of a psychophysical method designed to assign one unambiguous value to visual acuity, with non-arbitrary thresholds based on psychometric functions. Therefore, transforming "+/-" notation to actual results is in itself problematic and the below suggestion to convert it will remain an approximation to the most likely "true" result. Even more so, as the given conditions should work for charts with 4 or 5 optotypes in a line, and visual acuity is not always tested on such charts. Yet, I believe that the approach is still better than just omitting the letters or (worse) assigning a missing value to those entries.

See also